requirements.txt
django
djangorestframework
waitress
